scripts/requirements-py39.txt
beautifulsoup4>=4.14.2
bottle>=0.13.4
deathbycaptcha-official>=4.6.0
dukpy>=0.5.0
# Python 3.9 compatibility: Pillow 11/12 require >=3.10
pillow>=10.0.0,<11
pycryptodomex>=3.23.0
requests>=2.32.5
urllib3>=2.6.0
